requirements.txt
APScheduler==3.10.4
beautifulsoup4==4.12.3
certifi==2024.8.30
Flask==3.0.3
numpy==2.1.3
openai==0.28.0
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
pdf2image==1.17.0
Pillow==11.0.0
pymongo==4.10.1
PyPDF2==3.0.1
pytesseract==0.3.13
python-dotenv==1.0.1
Requests==2.32.3
selenium==4.26.1
webdriver_manager==4.0.2
